xfstests.py: Add support to specify custom output dir
Currently, avocado copies the output of xfstest run to self.outputdir,
which is a constant set by avocado and can't be changes. However there
are use cases where we want the output to be in a specific dir instead.

To allow this, add a new option in yaml named outputdir, that can be
used as something like:

    args: '-R xunit generic/001'
    mkfs_opt: '-b 4096'
    mount_opt: '-o block_validity'
    outputdir: '/mnt'

If this is provided, avocado would copy the xfstests result to
the provided dir else it would default to copying to self.outputdir

Signed-off-by: Ojaswin Mujoo <ojaswin@linux.ibm.com>